pyproject.toml
[project]
name = "swarm_hydra"
version = "0.0.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
license-files = ["LICEN[CS]E*"]
dependencies = [
    "einops==0.8.1",
    "h5py==3.12.1",
    "hydra-core==1.3.2",
    "imageio==2.36.1",
    "jinja2==3.1.6",
    "lint>=1.2.1",
    "matplotlib==3.10.0",
    "minisom==2.3.5",
    "numpy==1.26.4",
    "omegaconf==2.3.0",
    "ruff>=0.14.5",
    "scikit-learn==1.7.0",
    "scipy==1.15.2",
    "seaborn==0.13.2",
    "statsmodels>=0.14.5",
    "tqdm==4.67.1",
]

[build-system]
requires = ["setuptools >= 77.0.3"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
exclude = ["tests*", "docs*", "build*", "dist*", "*.egg-info"]
